Working basic implementation of the comp script
This only runs on Andrew's desktop (hard coded for
data paths) and does not actually report the correlation
(that isn't implemented yet) but it does (a) buld a reference
(e.g. tomographic) model and resolution then (b) read in comparison
models (e.g. from geodynamics) and reparameterises these before
(c) applying the reference filter to the comparison model and then
(d) calling the function to calculate the correlation.